requirements.txt
selenium
pytest
pytest-html
chromedriver-autoinstaller
pytesseract
PIL
pytest-xdist
python-dotenv